requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
python-multipart>=0.0.12
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Async
anyio>=4.6.0
httpx>=0.28.0
aiofiles>=24.0.0

# Redis Stack (vector search + KV)
redis[hiredis]>=5.2.0

# AI / LLM
anthropic>=0.40.0
openai>=1.55.0  # optional, for OpenAI embeddings

# Embeddings / ML
numpy>=1.26.0
sentence-transformers>=3.0.0
torch>=2.0.0

# Memory / Config
pyyaml>=6.0.2
click>=8.1.7
python-dotenv>=1.0.1

# Skills / Tools
requests>=2.32.0
beautifulsoup4>=4.12.0
playwright>=1.48.0

# Telegram
aiogram>=3.0.0

# Scheduling
apscheduler>=3.10.0

# Cryptography / Secrets
cryptography>=43.0.0
python-jose[cryptography]>=3.3.0

# Process management / utilities
psutil>=6.1.0
watchdog>=6.0.0

# Logging
structlog>=24.4.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
